requirements.txt
streamlit
fastapi
uvicorn
transformers
gtts
requests
beautifulsoup4
googlesearch-python
python-dotenv
torch
yake